Update Aiursoft.WebTools version and modify docker secrets loading
The version of Aiursoft.WebTools has been updated from 8.0.5 to 8.0.6. Additionally, the method for loading Docker secrets in the application was modified in the Extends.cs file. Instead of inserting Docker secrets at the beginning of the configuration sources, it now adds Docker secrets without assigning position, to simplify the configuration process.