Add trustAnyProxy option to SupportForwardHeadersPlugin
A new trustAnyProxy parameter was introduced to the SupportForwardHeadersPlugin constructor. This allows configuration of whether KnownNetworks and KnownProxies should be cleared or not. This is especially useful in Docker environments, where a reverse proxy is used and can drop all request headers and attach real IP, so we trust it. Consequently, the Aiursoft WebTools version was updated.